requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6

SQLAlchemy==2.0.36
asyncpg==0.30.0
alembic==1.13.3

pydantic==2.9.2
pydantic-settings==2.6.1

python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
pyjwt[crypto]==2.9.0

aiogram==3.13.1
redis==5.0.8

python-dateutil==2.9.0.post0
pytz==2024.1
